# exports/openapi.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/core-rules": {
      "get": {
        "description": "返回各核心规则集的最新版本，用于前端初始化展示。",
        "operationId": "CoreRulesController_listAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "核心规则集列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoreRuleSetVersionResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "列出所有核心规则集最新版本",
        "tags": [
          "CoreRules"
        ]
      }
    },
    "/core-rules/{coreName}/versions": {
      "get": {
        "description": "用于版本历史查询与回滚选择。",
        "operationId": "CoreRulesController_listByName",
        "parameters": [
          {
            "name": "coreName",
            "required": true,
            "in": "path",
            "description": "核心规则集名称",
            "schema": {
              "example": "core_sqli_rules",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "版本列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoreRuleSetVersionResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查看指定核心规则集的历史版本",
        "tags": [
          "CoreRules"
        ]
      },
      "post": {
        "description": "提交规则内容并生成新版本（从 v1.1 起可编辑）。",
        "operationId": "CoreRulesController_create",
        "parameters": [
          {
            "name": "coreName",
            "required": true,
            "in": "path",
            "description": "核心规则集名称",
            "schema": {
              "example": "core_sqli_rules",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCoreRuleSetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "新版本信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreRuleSetVersionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建核心规则集新版本",
        "tags": [
          "CoreRules"
        ]
      }
    },
    "/core-rules/{coreName}/versions/{versionNo}/rollback": {
      "post": {
        "description": "复制目标版本生成新版本，保留历史。",
        "operationId": "CoreRulesController_rollback",
        "parameters": [
          {
            "name": "coreName",
            "required": true,
            "in": "path",
            "description": "核心规则集名称",
            "schema": {
              "example": "core_sqli_rules",
              "type": "string"
            }
          },
          {
            "name": "versionNo",
            "required": true,
            "in": "path",
            "description": "目标版本号",
            "schema": {
              "example": 2,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackCoreRuleSetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "回滚后的新版本",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreRuleSetVersionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "回滚核心规则集到指定版本（生成新版本）",
        "tags": [
          "CoreRules"
        ]
      }
    },
    "/core-rules/{coreName}/restore-factory": {
      "post": {
        "description": "按出厂基线回滚并生成新版本。",
        "operationId": "CoreRulesController_restoreFactory",
        "parameters": [
          {
            "name": "coreName",
            "required": true,
            "in": "path",
            "description": "核心规则集名称",
            "schema": {
              "example": "core_sqli_rules",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackCoreRuleSetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "回滚后的新版本（出厂）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreRuleSetVersionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "恢复核心规则集到出厂版本 v1.0",
        "tags": [
          "CoreRules"
        ]
      }
    },
...
    "/logs/loki/waf/logs": {
      "get": {
        "description": "按时间范围与过滤条件查询审计日志。",
        "operationId": "LokiController_getWafLogs",
        "parameters": [
          {
            "name": "timeRange",
            "required": true,
            "in": "query",
            "description": "时间范围，如 5m/1h/24h",
            "schema": {
              "example": "1h",
              "type": "string"
            }
          },
          {
            "name": "server",
            "required": false,
            "in": "query",
            "description": "过滤的 serverName",
            "schema": {
              "example": "example.com",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "最大条数（1-1000）",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "example": 200,
              "type": "number"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "description": "过滤动作（ALLOW/BLOCK 等）",
            "schema": {
              "example": "BLOCK",
              "type": "string"
            }
          },
          {
            "name": "ruleId",
            "required": false,
            "in": "query",
            "description": "过滤规则 ID",
            "schema": {
              "example": "1001",
              "type": "string"
            }
          },
          {
            "name": "clientIp",
            "required": false,
            "in": "query",
            "description": "过滤客户端 IP",
            "schema": {
              "example": "1.2.3.4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Loki 查询结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询 WAF 审计日志（Loki）",
        "tags": [
          "Loki"
        ]
      }
    },
    "/logs/loki/waf/stats": {
      "get": {
        "description": "返回拦截数、攻击类型分布等统计聚合结果。",
        "operationId": "LokiController_getWafStats",
        "parameters": [
          {
            "name": "timeRange",
            "required": true,
            "in": "query",
            "description": "时间范围，如 5m/1h/24h",
            "schema": {
              "example": "1h",
              "type": "string"
            }
          },
          {
            "name": "server",
            "required": false,
            "in": "query",
            "description": "过滤的 serverName",
            "schema": {
              "example": "example.com",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "最大条数（1-1000）",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "example": 200,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统计结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取 WAF 统计（Loki 查询结果）",
        "tags": [
          "Loki"
        ]
      }
    },
    "/logs/loki/access/stats": {
      "get": {
        "description": "返回 PV/UV/QPS 等访问指标统计。",
        "operationId": "LokiController_getAccessStats",
        "parameters": [
          {
            "name": "timeRange",
            "required": true,
            "in": "query",
            "description": "时间范围，如 5m/1h/24h",
            "schema": {
              "example": "1h",
              "type": "string"
            }
          },
          {
            "name": "server",
            "required": false,
            "in": "query",
            "description": "过滤的 serverName",
            "schema": {
              "example": "example.com",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "最大条数（1-1000）",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "example": 200,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统计结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "访问统计（Loki 查询结果）",
        "tags": [
          "Loki"
        ]
      }
    },
    "/logs/loki/access/timeseries": {
      "get": {
        "description": "返回指定时间范围内的时序折线数据。",
        "operationId": "LokiController_getAccessTimeseries",
        "parameters": [
          {
            "name": "timeRange",
            "required": true,
            "in": "query",
            "description": "时间范围，如 5m/1h/24h",
            "schema": {
              "example": "1h",
              "type": "string"
            }
          },
          {
            "name": "server",
            "required": false,
            "in": "query",
            "description": "过滤的 serverName",
            "schema": {
              "example": "example.com",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "最大条数（1-1000）",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "example": 200,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "时序结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "访问时序（Loki 查询结果）",
        "tags": [
          "Loki"
        ]
      }
    },
    "/logs/loki/geo/world": {
      "get": {
        "description": "返回国家级访问/拦截地理聚合。",
        "operationId": "LokiController_geoWorld",
        "parameters": [
          {
            "name": "timeRange",
            "required": true,
            "in": "query",
            "description": "时间范围，如 5m/1h/24h",
            "schema": {
              "example": "1h",
              "type": "string"
            }
          },
          {
            "name": "server",
            "required": false,
            "in": "query",
            "description": "过滤的 serverName",
            "schema": {
              "example": "example.com",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "最大条数（1-1000）",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "example": 200,
              "type": "number"
            }
          },
          {
            "name": "mode",
            "required": true,
            "in": "query",
            "description": "模式：visit=访问量，block=拦截量",
            "schema": {
              "type": "string",
              "enum": [
                "visit",
                "block"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "地理聚合结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "世界地图聚合（Loki 查询结果）",
        "tags": [
          "Loki"
        ]
      }
    },
    "/logs/loki/geo/china": {
      "get": {
        "description": "返回省市级访问/拦截地理聚合。",
        "operationId": "LokiController_geoChina",
        "parameters": [
          {
            "name": "timeRange",
            "required": true,
            "in": "query",
            "description": "时间范围，如 5m/1h/24h",
            "schema": {
              "example": "1h",
              "type": "string"
            }
          },
          {
            "name": "server",
            "required": false,
            "in": "query",
            "description": "过滤的 serverName",
            "schema": {
              "example": "example.com",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "最大条数（1-1000）",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "example": 200,
              "type": "number"
            }
          },
          {
            "name": "mode",
            "required": true,
            "in": "query",
            "description": "模式：visit=访问量，block=拦截量",
            "schema": {
              "type": "string",
              "enum": [
                "visit",
                "block"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "地理聚合结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "中国地图聚合（Loki 查询结果）",
        "tags": [
          "Loki"
        ]
      }
    },
...
  },
  "info": {
    "title": "WAF 控制面 API",
    "description": "v1 接口文档",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CoreRuleSetVersionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "版本 ID",
            "example": 12
          },
          "coreName": {
            "type": "string",
            "description": "核心规则集名称",
            "example": "core_sqli_rules"
          },
          "versionNo": {
            "type": "number",
            "description": "版本号",
            "example": 3
          },
          "rulesJson": {
            "type": "object",
            "description": "核心规则 JSON",
            "additionalProperties": true
          },
          "status": {
            "type": "string",
            "description": "状态",
            "enum": [
              "SUCCESS",
              "FAILED",
              "ROLLED_BACK"
            ]
          },
          "note": {
            "type": "string",
            "description": "备注"
          },
          "createdBy": {
            "type": "string",
            "description": "创建人",
            "example": "admin"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间（ISO）"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间（ISO）"
          }
        },
        "required": [
          "id",
          "coreName",
          "versionNo",
          "rulesJson",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "RuleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "规则 ID（整数）",
            "example": 1001
          },
          "target": {
            "type": "string",
            "description": "匹配目标（可为单个或数组）",
            "example": "URI",
            "enum": [
              "URI",
              "ARGS_COMBINED",
              "BODY",
              "HEADER",
              "CLIENT_IP",
              "ALL_PARAMS"
            ]
          },
          "match": {
            "type": "string",
            "description": "匹配方式",
            "example": "CONTAINS",
            "enum": [
              "CONTAINS",
              "EXACT",
              "REGEX",
              "CIDR"
            ]
          },
          "pattern": {
            "type": "object",
            "description": "匹配模式（可为单个或数组）",
            "example": "/admin"
          },
          "action": {
            "type": "string",
            "description": "动作",
            "example": "DENY",
            "enum": [
              "DENY",
              "LOG",
              "BYPASS"
            ]
          },
          "score": {
            "type": "number",
            "description": "扣分（>=0，仅动态信誉场景使用）",
            "example": 5
          },
          "caseless": {
            "type": "boolean",
            "description": "忽略大小写",
            "example": true
          },
          "negate": {
            "type": "boolean",
            "description": "取反（negate）",
            "example": false
          },
          "headerName": {
            "type": "string",
            "description": "当 target=HEADER 时必填，指定 header 名称",
            "example": "User-Agent"
          },
          "tags": {
            "description": "规则标签（用于攻击类型归类）",
            "example": [
              "sqli"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "target",
          "match",
          "pattern",
          "action"
        ]
      },
      "CreateCoreRuleSetDto": {
        "type": "object",
        "properties": {
          "coreName": {
            "type": "string",
            "description": "核心规则集名称",
            "example": "core_sqli_rules"
          },
          "rules": {
            "description": "规则列表（friendly JSON 结构）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleDto"
            }
          },
          "note": {
            "type": "string",
            "description": "备注",
            "example": "补充 SQLi 规则"
          },
          "actor": {
            "type": "string",
            "description": "操作者",
            "example": "alice"
          }
        },
        "required": [
          "coreName",
          "rules"
        ]
      },
      "RollbackCoreRuleSetDto": {
        "type": "object",
        "properties": {
          "actor": {
            "type": "string",
            "description": "操作者",
            "example": "alice"
          }
        }
      },
...
    }
  }
}